src/lambda_functions/auth_authorizer/requirements.txt
boto3>=1.28.0
pyjwt>=2.6.0
cryptography>=38.0.0